Fix play method and create process_play method
The checking of the letter against what is on the board was taken out of
the play method and put into the new process_play method.  This was done
because the play method was getting too big.

The reduce has been changed to a destructive map in order for the
changes to stick.  Also had to include an unless conditonal in order for
the correct boolean to behave correctly; letters that were previously
found and already had their :visible set to true would always set
correct to true.