Make a string version of the secret word
The secret word is now displayed after the player loses the game.  This
was made easier by having a simple string copy of the secret word.  The
secret_letters is an array of hashes that has to determine whether to
show each letter to the player or not.  It's easier to just make a
string copy instead of trying to convert the secret_letters.