Move upcase from process_play to play
There was an issue with how the game handled the case of the letter.  It
was not converting the case to upcase till later in the process, which
threw off the duplicate letter check.

The upcase is not placed before letter_valid? check in play because at
this point, it is not known yet what exactly letter is; it could be nil
and trying to do an upcase on that would cause an error.